dbeaver/dbeaver-ee#980 create user or other entity with password base dialog
