+++ plugins/org.jkiss.dbeaver.ext.mssql.ui/src/org/jkiss/dbeaver/ext/mssql/ui/SQLServerUIMessages.properties
@@ -19,6 +19,3 @@ dialog_create_db_label_db_name=Database name
 dialog_create_db_title=Create database
 
 dialog_create_login_shell_title = Create login
-dialog_create_login_group = Settings
-dialog_create_login_label_name = Name
-dialog_create_login_label_password = Password
