+++ plugins/org.jkiss.dbeaver.ext.mssql.ui/src/org/jkiss/dbeaver/ext/mssql/ui/SQLServerUIMessages_de.properties
@@ -21,6 +21,3 @@ dialog_create_db_label_db_name = Name der Datenbank
 dialog_create_db_title = Datenbank erstellen
 
 dialog_create_login_shell_title = Login erstellen
-dialog_create_login_group = Einstellungen
-dialog_create_login_label_name = Name
-dialog_create_login_label_password = Passwort
