+++ plugins/org.jkiss.dbeaver.ext.mssql.ui/src/org/jkiss/dbeaver/ext/mssql/ui/SQLServerUIMessages_ru.properties
@@ -18,6 +18,3 @@ dialog_create_db_title=\u0421\u043E\u0437\u0434\u0430\u0442\u044C \u0431\u0430\u
 dialog_create_db_label_db_name=\u041D\u0430\u0437\u0432\u0430\u043D\u0438\u0435 \u0431\u0430\u0437\u044B \u0434\u0430\u043D\u043D\u044B\u0445
 
 dialog_create_login_shell_title = \u0421\u043E\u0437\u0434\u0430\u0442\u044C \u041B\u043E\u0433\u0438\u043D
-dialog_create_login_group = \u041D\u0430\u0441\u0442\u0440\u043E\u0439\u043A\u0438
-dialog_create_login_label_name = \u0418\u043C\u044F
-dialog_create_login_label_password = \u041F\u0430\u0440\u043E\u043B\u044C
