+++ plugins/org.jkiss.dbeaver.ui/src/org/jkiss/dbeaver/ui/internal/UIMessages.properties
@@ -43,6 +43,10 @@ controls_locale_selector_label_variant = Variant
 
 edit_text_dialog_title_edit_value=Edit value
 
+dialogs_name_and_password_dialog_group_settings = Settings
+dialogs_name_and_password_dialog_label_name = Name
+dialogs_name_and_password_dialog_label_password = Password
+
 ui_properties_name = Name
 ui_properties_value = Value
