+++ plugins/org.jkiss.dbeaver.ui/src/org/jkiss/dbeaver/ui/internal/UIMessages_de.properties
@@ -40,3 +40,7 @@ ui_properties_name = Name
 ui_properties_value = Wert
 
 control_label_connection_folder = Verbindungsordner
+
+dialogs_name_and_password_dialog_group_settings = Einstellungen
+dialogs_name_and_password_dialog_label_name = Name
+dialogs_name_and_password_dialog_label_password = Passwort
