+++ plugins/org.jkiss.dbeaver.ui/src/org/jkiss/dbeaver/ui/internal/UIMessages_ru.properties
@@ -39,6 +39,10 @@ controls_locale_selector_label_variant=\u0412\u0430\u0440\u0438\u0430\u043D\u044
 
 edit_text_dialog_title_edit_value=\u0418\u0437\u043C\u0435\u043D\u0438\u0442\u044C \u0437\u043D\u0430\u0447\u0435\u043D\u0438\u0435
 
+dialogs_name_and_password_dialog_group_settings = \u041D\u0430\u0441\u0442\u0440\u043E\u0439\u043A\u0438
+dialogs_name_and_password_dialog_label_name = \u0418\u043C\u044F
+dialogs_name_and_password_dialog_label_password = \u041F\u0430\u0440\u043E\u043B\u044C
+
 ui_properties_name=\u0418\u043C\u044F
 ui_properties_value=\u0417\u043D\u0430\u0447\u0435\u043D\u0438\u0435
